requirements.txt
discord.py
disputils
jikanpy
pytz
requests

